Docker file added. Path appending.
