+++ practice-app/DockerFile
@@ -1,0 +1,13 @@
+# syntax=docker/dockerfile:1
+
+FROM python:3.8-slim-buster
+
+WORKDIR /app
+
+COPY requirements.txt requirements.txt
+RUN pip3 install -r requirements.txt
+
+COPY sqlfiles sqlfiles
+COPY src src
+
+CMD [ "python3", "src/main/main.py"]
